test: consolidate recorder transport coverage
